finised serialising

+++ src/main/java/Essentials/Implementations/summTransaction.java
@@ -1,8 +1,9 @@
 package Essentials.Implementations;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 public class summTransaction implements Serializable {
-    public Integer rubles;
+    public BigDecimal rubles;
     public Integer cents;
 }
